fix for plaits chords engine + small fixes

+++ eurorack/braids/signature_waveshaper.h
@@ -64,8 +64,11 @@ class SignatureWaveshaper {
       int16_t x_skew = i * i - 32768;
       x = stmlib::Mix(x, x_skew, skew << 11);
       
-      int16_t sigmoid = x * (8192 + (sigmoid_strength << 10)) / \
-          (8192 + (sigmoid_strength * abs(x) >> 5));
+        // vb, signed integer overflow: this might not be handled gracefully by all compilers
+//      int16_t sigmoid = x * (8192 + (sigmoid_strength << 10)) / \
+//          (8192 + (sigmoid_strength * abs(x) >> 5));
+        int16_t sigmoid = x * (8192 + (sigmoid_strength << 10)) / \
+        (8192 + (int32_t)((sigmoid_strength * abs(x)) >> 5));
       int16_t bumplets = wav_sine[(i * bumplets_frequency) & 255];
       uint16_t bumplet_gain = x * x / (bumplets_width) + 16;
       bumplet_gain = 32768 * 128 / (128 + bumplet_gain);
